Add initial firecloud-project template and schema.
This commit captures a snapshot where it's possible to create a DM composite
type and deploy a simple project successfully. APIs are enabled in batches using
the Service Usage API, which speeds up the overall process quite a bit.

Currently the 'highSecurityNetwork' option is still not working due to some DM
template expansion errors. A subsequent commit will bring that option into
operation.