Add a top-level firecloud-project template and schema.
This template combines the project, network, and firewall templates into a template that can provision a complete FireCloud project, with Pub/Sub notifications at the start and end of provisioning.

There are a few ugly parts, mostly to work around the issues that one cannot use a template-defined Deployment Manager node in a subsequent "dependsOn" stanza. Instead, we adopt a pattern where each template explicitly outputs a 'resourceNames' property, which can be referred to when a later node needs to be blocked by the template node.

A future commit will add a README with example commands for (1) storing the firecloud-project template as a composite type and (2) instantiating a new firecloud-project deployment.