Distinguish between resource name and GCP object name for net/subnets.
This fixes an issue where Deployment Manager isn't happy if we create
subnets all with the same name. We'd previously been using the single
"name" property as the DM resource name as well as the GCP object
name; by separating them, we can keep DM happy with globally unique
resource names while keeping the GCP names identical.

Technically this change only needed to be made for subnets, but I made
it work the same in our network template for consistency.